+++ src/gt4py/next/ffront/field_ast_to_sdfg/foast_to_sdfg.py
@@ -1,0 +1,10 @@
+import dace
+
+import gt4py.eve as eve
+from gt4py.next.ffront import program_ast as past
+from gt4py.next.type_system import type_specifications as ts
+
+
+
+class PastToSDFG(eve.NodeVisitor):
+    sdfg: dace.SDFG
