+++ src/gt4py/next/ffront/field_ast_to_sdfg/utility.py
@@ -1,0 +1,16 @@
+from gt4py.next.type_system import type_specifications as ts
+import dace
+
+
+def type_spec_to_dtype(type_: ts.ScalarType):
+    if type_.kind == ts.ScalarKind.BOOL:
+        return dace.bool_
+    elif type_.kind == ts.ScalarKind.INT32:
+        return dace.int32
+    elif type_.kind == ts.ScalarKind.INT64:
+        return dace.int64
+    elif type_.kind == ts.ScalarKind.FLOAT32:
+        return dace.float32
+    elif type_.kind == ts.ScalarKind.FLOAT64:
+        return dace.float64
+    raise ValueError(f"scalar type {type_} not supported")
