+++ tests/dace_tests/test_dace_backend.py
@@ -1,9 +1,9 @@
 import numpy as np
 
-from functional.common import Dimension, Field
-from functional.ffront.decorator import field_operator
-from functional.iterator.embedded import np_as_located_field
-from functional.program_processors.runners.dace_fieldview import run_dace_fieldview
+from gt4py.next.common import Dimension, Field
+from gt4py.next.ffront.decorator import field_operator
+from gt4py.next.iterator.embedded import np_as_located_field
+from gt4py.next.program_processors.runners.dace_fieldview import run_dace_fieldview
 
 
 IDim = Dimension("IDim")
